// frontend/src/api/generated/models/searchResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Dekopen API
 * Authenticated tenant and engine API boundary.
 * OpenAPI spec version: 1.0.0
 */

export interface SearchResult {
  group: string;
  id: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  path: string;
}